Last bit of refactoring I promise, all validation stuff is now grouped together
