--- service/src/main/java/uk/gov/hmcts/reform/fpl/interfaces/EPOGroup.java
@@ -1,3 +1,0 @@
-package uk.gov.hmcts.reform.fpl.interfaces;
-
-public @interface EPOGroup { }
