+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/GroundsForEPO.java
@@ -4,10 +4,9 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.fpl.interfaces.EPOGroup;
+import uk.gov.hmcts.reform.fpl.validation.groups.EPOGroup;
 
 import java.util.List;
-
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
