+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/groups/EPOGroup.java
@@ -1,0 +1,3 @@
+package uk.gov.hmcts.reform.fpl.validation.groups;
+
+public @interface EPOGroup { }
