+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/groups/NoticeOfProceedingsGroup.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.interfaces;
+package uk.gov.hmcts.reform.fpl.validation.groups;
 
 // JSR Validation interface used to validate the notice of proceedings case event
 public @interface NoticeOfProceedingsGroup {
