+++ service/src/main/java/uk/gov/hmcts/reform/fpl/validation/groups/UploadDocumentsGroup.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.interfaces;
+package uk.gov.hmcts.reform.fpl.validation.groups;
 
 // JSR Validation interface used to validate the upload documents case event
 public @interface UploadDocumentsGroup {
