+++ service/src/test/java/uk/gov/hmcts/reform/fpl/validation/validators/documents/HasAttachedDocumentValidatorTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.hmcts.reform.fpl.interfaces.UploadDocumentsGroup;
 import uk.gov.hmcts.reform.fpl.model.common.Document;
 import uk.gov.hmcts.reform.fpl.service.ValidateGroupService;
+import uk.gov.hmcts.reform.fpl.validation.groups.UploadDocumentsGroup;
 
 import java.util.List;
 import javax.validation.Validation;
